# src/app/api/programs/custody/run/route.ts
// ---------------------------------------------------------------------------
// POST /api/programs/custody/run — CUSTODY verify-bundle endpoint (DEPRECATED ALIAS)
// ---------------------------------------------------------------------------
//
// !! DEPRECATED — clients should POST to /api/v1/runs with
//    { pipeline: "bureau:custody", payload: { ...this body... } }. !!
//
// Wave-2 migration: this route stays alive as a deprecated alias so callers
// that haven't migrated keep working, but it now validates via the shared
// `validateCustodyPayload` and dual-writes into the v1 store so legacy
// + /v1/runs callers converge on the same phraseId for the same payload.
//
// Day-1 contract preserved:
//   1. Auth check by Supabase JWT cookie presence.
//   2. CSRF defence — same-origin enforced.
//   3. https-only bundleUrl + private-IP block + optional expectedVendor
//      hostname grammar + ToS / authorization assertion.
//   4. On success: { runId, phraseId, bundleUrl, expectedVendor (or null),
//      status:"verification pending", deprecated: true,
//      replacement: "/api/v1/runs" } + RFC 8594 Deprecation/Sunset/Link
//      headers. runId === phraseId — single primitive, identical on
//      idempotent retries (mirrors DRAGNET M5).
//
// Idempotency contract: legacy CUSTODY callers double-clicking within
// ~60s dedupe to the SAME phraseId as a /v1/runs caller posting the same
// body — we synthesize the same minute-bucketed key the RunForm uses
// (`custody:<vendorOrUnknown>:<bundleUrl>:<minute>`) before delegating.
// ---------------------------------------------------------------------------

import { NextResponse } from "next/server";

import {
  isAuthed,
  isSameSiteRequest,
  rateLimitOk,
} from "../../../../../lib/security/request-guards";
import { validateCustodyPayload } from "../../../../../lib/v1/pipeline-validators";
import { createRun } from "../../../../../lib/v1/run-store";

const DEPRECATION_HEADERS: Record<string, string> = {
  Deprecation: "true",
  Sunset: "Wed, 31 Dec 2026 23:59:59 GMT",
  Link: '</api/v1/runs>; rel="successor-version"',
};

interface CustodyRequestBody {
  bundleUrl?: string;
  expectedVendor?: string;
  authorizationAcknowledged?: boolean;
}

/**
 * Synthesize the same minute-bucketed idempotency key the CUSTODY
 * RunForm sends to /v1/runs. Format:
 *   `custody:<vendorOrUnknown>:<bundleUrl>:<minute-bucket>`
 *
 * `vendorOrUnknown` is the operator-supplied expectedVendor when present,
 * else the literal string "unknown" (so a generic bundle still has a
 * dedupe key shape that's stable across the legacy + /v1/runs surfaces).
 *
 * Legacy double-click + /v1/runs double-click with the same payload land
 * on the SAME stored run record. Mirrors DRAGNET's C1 fix.
 */
function synthesizeIdempotencyKey(
  vendorOrUnknown: string,
  bundleUrl: string,
  now: number = Date.now(),
): string {
  const minuteBucket = Math.floor(now / 60_000);

  return `custody:${vendorOrUnknown}:${bundleUrl}:${minuteBucket}`;
}

export async function POST(req: Request): Promise<Response> {
  if (!isSameSiteRequest(req)) {
    return NextResponse.json(
      { error: "cross-site request rejected" },
      { status: 403 },
    );
  }
  if (!rateLimitOk(req)) {
    return NextResponse.json(
      { error: "too many requests — slow down and try again in a minute" },
      { status: 429 },
    );
  }
  if (!isAuthed(req)) {
    return NextResponse.json(
      {
        error: "authentication required",
        signInUrl: "/sign-in?redirect=/programs/custody/run",
      },
      { status: 401 },
    );
  }
  let body: CustodyRequestBody;
  try {
    body = (await req.json()) as CustodyRequestBody;
  } catch {
    return NextResponse.json(
      { error: "invalid JSON body" },
      { status: 400 },
    );
  }

  // Single source of truth — the same validator /v1/runs uses. Keeps the
  // two surfaces from drifting.
  const result = validateCustodyPayload(body);
  if (!result.ok) {
    return NextResponse.json({ error: result.error }, { status: 400 });
  }

  // After validation passes, normalize for the response shape + dedupe key.
  const bundleUrl = (body.bundleUrl ?? "").trim();
  const explicitVendor = (body.expectedVendor ?? "").trim().toLowerCase();
  const vendorOrUnknown =
    explicitVendor.length > 0 ? explicitVendor : "unknown";

  // Delegate persistence to the unified /v1/runs store so the receipt
  // page can read this run back via GET /api/v1/runs/[id]. The store
  // assigns the canonical vendor- or bundle-scoped phraseId — that
  // becomes the user-facing runId.
  //
  // CUSTODY's vendor-scoped vs. bundleUrl-scoped phrase-prefix logic
  // lives in run-store#runIdForProgram: when expectedVendor is supplied,
  // we promote it to vendorDomain for `generateScopedPhraseId`; else
  // the bundle hostname is used. Form must omit `expectedVendor` when
  // empty so canonicalJson skips the field and legacy + /v1/runs hash
  // identically.
  const { record } = createRun({
    pipeline: "bureau:custody",
    payload: {
      bundleUrl,
      // expectedVendor is canonical; promote to vendorDomain so the
      // store's scoping prefers expectedVendor over the bundle hostname.
      vendorDomain: explicitVendor.length > 0 ? explicitVendor : undefined,
      expectedVendor: explicitVendor.length > 0 ? explicitVendor : undefined,
      authorizationAcknowledged: body.authorizationAcknowledged,
    },
    idempotencyKey: synthesizeIdempotencyKey(vendorOrUnknown, bundleUrl),
  });

  return NextResponse.json(
    {
      runId: record.runId,
      phraseId: record.runId,
      bundleUrl,
      expectedVendor: explicitVendor.length > 0 ? explicitVendor : null,
      status: "verification pending",
      deprecated: true,
      replacement: "/api/v1/runs",
      note: "deprecated alias — POST to /api/v1/runs with pipeline=bureau:custody",
    },
    { status: 200, headers: DEPRECATION_HEADERS },
  );
}